Move toasts to layout folder
