+++ frontend/src/dashboard/FileUpload.tsx
@@ -1,6 +1,6 @@
 import {useState, useRef} from "react";
 import {Button} from "../components/Button";
-import {useShowToast} from "../components/toastStore.ts";
+import {useShowToast} from "../layout/toastStore.ts";
 
 function FileUploader({refreshFileList}: {refreshFileList: () => void}) {
 	const [fileUploads, setFileUploads] = useState<FileList | null>(null);
