+++ frontend/src/user/layout/TopBar.tsx
@@ -1,6 +1,6 @@
 import {Link, useNavigate} from "react-router";
 import {Button} from "../../components/Button";
-import {useShowToast} from "../../components/toastStore.ts";
+import {useShowToast} from "../../layout/toastStore.ts";
 import type {MouseEvent} from "react";
 
 export default function TopBar() {
